# business/docs/swagger.yaml
basePath: /api/v1
definitions:
  calls.RecordingConfig:
    properties:
      ackAt:
        type: string
      ackBy:
        type: integer
      available:
        description: |-
          Available é falso quando a instalação não tem armazenamento de objetos: as
          opções de gravação não são oferecidas.
        type: boolean
      mode:
        type: string
    type: object
...
  controllers.recordingConfigRequest:
    properties:
      ack:
        description: Ack confirma que o administrador leu e aceitou o termo de responsabilidade.
        type: boolean
      mode:
        type: string
    required:
    - mode
    type: object
...
host: localhost:8082
info:
  contact:
    email: dev@watink.com.br
    name: Watink Dev
  description: API do backend Go — atendimento e automação WhatsApp.
  license:
    name: Proprietary
    url: https://watink.com.br
  termsOfService: https://watink.com.br/terms
  title: Watink Business API
  version: 2.0.0
paths:
...
  /calls/{id}/audio:
    get:
      description: Abre o canal de áudio PCM 16 kHz mono Int16 (quadros de 640 bytes)
        do operador que assumiu a chamada. Auth por token na query, como o SSE.
      parameters:
      - description: ID da chamada
        in: path
        name: id
        required: true
        type: string
      - description: JWT
        in: query
        name: token
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: Áudio da chamada (WebSocket)
      tags:
      - calls
...
  /calls/{id}/recording:
    delete:
      parameters:
      - description: ID da chamada
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Excluir a gravação
      tags:
      - calls
    get:
      description: Devolve uma URL assinada e temporária e registra quem ouviu. Fora
        do alcance do usuário responde 404.
      parameters:
      - description: ID da chamada
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ouvir a gravação
      tags:
      - calls
  /calls/{id}/recording/start:
    post:
      parameters:
      - description: ID da chamada
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Iniciar gravação da chamada (modo opcional)
      tags:
      - calls
  /calls/{id}/recording/stop:
    post:
      parameters:
      - description: ID da chamada
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Parar a gravação da chamada (modo opcional)
      tags:
      - calls
...
  /calls/recording-config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calls.RecordingConfig'
      security:
      - BearerAuth: []
      summary: Configuração de gravação de chamadas
      tags:
      - calls
    put:
      consumes:
      - application/json
      description: Sair de "off" exige ack=true (aceite do termo); o aceite grava
        usuário e horário.
      parameters:
      - description: Modo e aceite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.recordingConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calls.RecordingConfig'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Alterar o modo de gravação de chamadas
      tags:
      - calls
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"